requirements.txt
telethon==1.32.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
aiosqlite==0.19.0
cryptg==0.4.0 